optimize cloneset event handler codes
